Renamed TransferProxy to Proxy
